Recommendar/requirements.txt
flask
pandas
numpy
joblib
sentence-transformers
schedule
pymongo
scikit-learn
requests
python-dotenv